refactor(tests): Reorg tests into actions/versions folders.
