+++ test/addLiquidity/v2/addLiquidityGyro3.integration.test.ts
@@ -26,15 +26,15 @@ import {
     getPoolAddress,
     AddLiquidityUnbalancedInput,
     AddLiquiditySingleTokenInput,
-} from '../src';
-import { forkSetup } from './lib/utils/helper';
+} from '../../../src';
+import { forkSetup } from '../../lib/utils/helper';
 import {
     assertAddLiquidityProportional,
     doAddLiquidity,
-} from './lib/utils/addLiquidityHelper';
-import { AddLiquidityTxInput } from './lib/utils/types';
-import { ANVIL_NETWORKS, startFork } from './anvil/anvil-global-setup';
-import { addLiquidityKindNotSupportedByGyro } from '../src/entities/addLiquidity/utils/validateInputs';
+} from '../../lib/utils/addLiquidityHelper';
+import { AddLiquidityTxInput } from '../../lib/utils/types';
+import { ANVIL_NETWORKS, startFork } from '../../anvil/anvil-global-setup';
+import { addLiquidityKindNotSupportedByGyro } from '../../../src/entities/addLiquidity/utils/validateInputs';
 
 const { rpcUrl } = await startFork(ANVIL_NETWORKS.POLYGON);
 const chainId = ChainId.POLYGON;
