+++ test/removeLiquidity/v2/removeLiquidityComposableStable.integration.test.ts
@@ -26,16 +26,16 @@ import {
     getPoolAddress,
     RemoveLiquidityInput,
     InputAmount,
-} from '../src';
-import { forkSetup } from './lib/utils/helper';
-import { RemoveLiquidityTxInput } from './lib/utils/types';
+} from '../../../src';
+import { forkSetup } from '../../lib/utils/helper';
+import { RemoveLiquidityTxInput } from '../../lib/utils/types';
 import {
     assertRemoveLiquidityProportional,
     assertRemoveLiquiditySingleToken,
     assertRemoveLiquidityUnbalanced,
     doRemoveLiquidity,
-} from './lib/utils/removeLiquidityHelper';
-import { ANVIL_NETWORKS, startFork } from './anvil/anvil-global-setup';
+} from '../../lib/utils/removeLiquidityHelper';
+import { ANVIL_NETWORKS, startFork } from '../../anvil/anvil-global-setup';
 
 const chainId = ChainId.MAINNET;
 const { rpcUrl } = await startFork(ANVIL_NETWORKS.MAINNET);
