+++ test/swaps/gyro3Pool.test.ts
@@ -1,7 +1,7 @@
 // pnpm test -- gyro3Pool.test.ts
-import testPools from './lib/testData/testPools/gyro3_44133130.json';
-import { ChainId, RawGyro3Pool, SwapKind, Token, TokenAmount } from '../src';
-import { Gyro3Pool } from '../src/entities/pools/gyro3';
+import testPools from '../lib/testData/testPools/gyro3_44133130.json';
+import { ChainId, RawGyro3Pool, SwapKind, Token, TokenAmount } from '../../src';
+import { Gyro3Pool } from '../../src/entities/pools/gyro3';
 
 describe('Gyro3Pool tests USDC > DAI', () => {
     const testPool = { ...testPools }.pools[0] as RawGyro3Pool;
