+++ test/swaps/gyroEPool.test.ts
@@ -1,8 +1,8 @@
 // pnpm test -- test/gyroEPool.test.ts
-import testPools from './lib/testData/testPools/gyroE_44215395.json';
-import { ChainId, SwapKind, Token, TokenAmount } from '../src';
-import { RawGyroEPool } from '../src/data/types';
-import { GyroEPool } from '../src/entities/pools/gyroE';
+import testPools from '../lib/testData/testPools/gyroE_44215395.json';
+import { ChainId, SwapKind, Token, TokenAmount } from '../../src';
+import { RawGyroEPool } from '../../src/data/types';
+import { GyroEPool } from '../../src/entities/pools/gyroE';
 
 describe('gyroEPool tests', () => {
     const testPool = { ...testPools }.pools[0] as RawGyroEPool;
