+++ test/swaps/gyroEV2Pool.test.ts
@@ -1,7 +1,7 @@
 // pnpm test -- gyroEV2Pool.test.ts
-import { ChainId, RawGyroEPool, SwapKind, Token, TokenAmount } from '../src';
-import { GyroEPool } from '../src/entities/pools/gyroE';
-import testPools from './lib/testData/testPools/gyroE_44215395.json';
+import { ChainId, RawGyroEPool, SwapKind, Token, TokenAmount } from '../../src';
+import { GyroEPool } from '../../src/entities/pools/gyroE';
+import testPools from '../lib/testData/testPools/gyroE_44215395.json';
 
 describe('gyroEPool tests', () => {
     const testPool = { ...testPools }.pools[1] as RawGyroEPool;
