+++ test/swaps/weightedPool.integration.test.ts
@@ -2,18 +2,18 @@
 import dotenv from 'dotenv';
 dotenv.config();
 
-import { SmartOrderRouter } from '../src/sor';
-import { sorGetSwapsWithPools } from '../src/static';
-import { ChainId, ETH, BATCHSIZE, VAULT } from '../src/utils';
-import { Token, TokenAmount } from '../src/entities';
-import { OnChainPoolDataEnricher } from '../src/data/enrichers/onChainPoolDataEnricher';
-import { SwapKind, SwapOptions } from '../src/types';
-import { BasePool } from '../src/entities/pools';
-import { MockPoolProvider } from './lib/utils/mockPoolProvider';
+import { SmartOrderRouter } from '../../src/sor';
+import { sorGetSwapsWithPools } from '../../src/static';
+import { ChainId, ETH, BATCHSIZE, VAULT } from '../../src/utils';
+import { Token, TokenAmount } from '../../src/entities';
+import { OnChainPoolDataEnricher } from '../../src/data/enrichers/onChainPoolDataEnricher';
+import { SwapKind, SwapOptions } from '../../src/types';
+import { BasePool } from '../../src/entities/pools';
+import { MockPoolProvider } from '../lib/utils/mockPoolProvider';
 
-import testPools from './lib/testData/testPools/weighted_17473810.json';
-import { RawWeightedPool } from '../src';
-import { ANVIL_NETWORKS, startFork } from './anvil/anvil-global-setup';
+import testPools from '../lib/testData/testPools/weighted_17473810.json';
+import { RawWeightedPool } from '../../src';
+import { ANVIL_NETWORKS, startFork } from '../anvil/anvil-global-setup';
 
 const chainId = ChainId.MAINNET;
 const { rpcUrl } = await startFork(ANVIL_NETWORKS.MAINNET);
